requirements.txt
opencv-python
torch
ultralytics
numpy
matplotlib
Pillow
